# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Docs",
        "description": "Dokumentasi API untuk aplikasi mobile Himakom",
        "version": "1.0.0"
    },
    "paths": {
        "/api/cabinets": {
            "get": {
                "tags": [
                    "Cabinets"
                ],
                "summary": "GET cabinets",
                "description": "Return data departments yang akan ditampilkan pada halaman cabinets mobile",
                "operationId": "a0bb0ac7fe50ff6e3cdd3988523b7d49",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/departments": {
            "get": {
                "tags": [
                    "Departments"
                ],
                "summary": "GET departments",
                "description": "Return data departments yang akan ditampilkan pada halaman departemen mobile",
                "operationId": "609fe4b236597615a801133f5a5c618d",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "GET events",
                "description": "Return data events yang akan ditampilkan pada halaman events mobile",
                "operationId": "02564acadf6f48e7cf453bb2319ac1b4",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "GET users",
                "description": "Return data users yang akan ditampilkan pada halaman dashboard mobile",
                "operationId": "c5ccc940808f87add4da9d1db1317605",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "GET user",
                "description": "Return data user yang akan ditampilkan pada halaman profile mobile",
                "operationId": "9142ec228ea53995c67de772a779eed5",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/device-token": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update device token",
                "description": "Update device token",
                "operationId": "5db00c1b9ed75a0640013bd589329ca8",
                "requestBody": {
                    "description": "Update device token",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "device_token"
                                ],
                                "properties": {
                                    "device_token": {
                                        "type": "string",
                                        "example": "device_token"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/loginApi": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Autentikasi user dan mengembalikan token",
                "operationId": "5a8b60d1273fdaa798460c9a585719f5",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/logoutApi": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout",
                "description": "Logout user",
                "operationId": "60739dcf4867e33ca373fbc907872c32",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "Authorization Control",
                "scheme": "bearer"
            }
        }
    }
}